optimized rendering for selection and camera animation for web and mobile [mobile + web + libs/hex + libs/ui]
