+++ apps/web/src/pages/Home.tsx
@@ -1,11 +1,11 @@
-import { Home as H, Planet } from '@end/components';
+import { getPointInBetweenByPerc, Home as H } from '@end/components';
 import { database, sync } from '@end/wm/web';
-import React, { useMemo, useRef, useState } from 'react';
+import React, { useCallback, useMemo, useRef, useState } from 'react';
 import { Canvas } from '@react-three/fiber';
 import * as THREE from 'three';
-import { useHexasphere } from '@end/hexasphere';
 import { useWindowDimensions } from 'react-native';
 import { OrbitControls } from '@react-three/drei';
+import {  hexasphere, Hexasphere } from '@end/hexasphere';
 
 export default function Home() {
   const ref = useRef(null);
@@ -41,50 +41,19 @@ export default function Home() {
 
     return cam;
   }, []);
-  const { tiles, hexasphere, setReset, reset } = useHexasphere();
-  const [selectedTile, setSelectedTile] = useState<{
-    x: number;
-    y: number;
-    z: number;
-  }>();
 
   return (
     <H database={database} sync={sync} apiUrl={process.env.API_BASE_URL}>
-      <Planet
-        setReset={setReset}
-        reset={reset}
-        tiles={tiles}
-        hexasphere={hexasphere}
-        selectedTile={selectedTile}
-        setSelectedTile={(id) =>
-          setSelectedTile(
-            (prevId: { x: number; y: number; z: number } | undefined) => {
-              const { x, y, z } = prevId ?? {};
-              const { x: x1, y: y1, z: z1 } = id ?? {};
-              return JSON.stringify({ x, y, z }) === JSON.stringify({ x: x1, y: y1, z: z1 })
-                ? undefined
-                : id;
-            }
-          )
-        }
+      <Canvas
+        style={{
+          flex: 1,
+          ...responsiveness,
+        }}
+        camera={cam}
       >
-        {(hexasphere, controls, footer) => (
-          <>
-            <Canvas
-              style={{
-                flex: 1,
-                ...responsiveness,
-              }}
-              camera={cam}
-            >
-              <OrbitControls />
-              {hexasphere}
-            </Canvas>
-            {controls}
-            {footer}
-          </>
-        )}
-      </Planet>
+        <Hexasphere />
+        <OrbitControls />
+      </Canvas>
     </H>
   );
 }
